Refactored Flask app to use blueprint, split routes code into several different files under routes/*
