+++ templates/page.html
@@ -113,7 +113,7 @@
          <li>
             <span class="timestamp">{{ sa.ts_text }}</span>
             <img src="{{ url_for('static', filename = 'sources/' + sa.domain + '.png') }}" width="16" height="16">
-            <a href="{{ url_for('page', id = sa.uuid) }}"><span class="heading">{{ sa.heading }}</span></a>
+            <a href="{{ url_for('routes.page', id = sa.uuid) }}"><span class="heading">{{ sa.heading }}</span></a>
             <span class="similarity">{{ sa.similarity | format_is(1) }}%</span>
          </li>
 {% endfor %}
