+++ templates/test.html
@@ -17,7 +17,7 @@
 
    <div class="input">
 
-      <p class="header"><a href="{{ url_for('main') }}"><img src="{{ url_for('static', filename='img/GreynirLogo404x180.png') }}" width=202 height=90></a></p>
+      <p class="header"><a href="{{ url_for('routes.main') }}"><img src="{{ url_for('static', filename='img/GreynirLogo404x180.png') }}" width=202 height=90></a></p>
 
       <p class="author">málgreinir fyrir íslensku<br>
       <b>alfa-prófunarútgáfa</b><br>
